Use name() instead of toString() for enum value.
The javadoc for Enum notes that name() should be used when
correctness depends on getting the exact name of the enum
constant which will not vary from release to release. The
javadoc further encourages overriding the toString() method
when appropriate to provide a more user-friendly name. This
commit fixes the DefaultEnumConverterFactory in cases where
the toString() method is overridden to return something other
than the enum constant.