Skip bridge-methods for POJO-getter/setter #314
Bridge-methods are generated by the compiler. In case of #314, getter and setter are created because the class inherited from a parametr. class and the getter/setter work on Object (as in the parent-class, after type erasure). That's why they need to be excluded from the list of methods to choose getters and setters from.